third_party/blink/web_tests/fast/events/event-fire-disconnected-bubbling-mutation-expected.txt

Mutation events should propagate in disconnected DOM trees.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS eventReceivedOnParent is true
PASS eventReceivedOnChild is true
PASS successfullyParsed is true

TEST COMPLETE
